@@ -7,7 +7,6 @@ members = [
   "neqo-server",
   "neqo-qpack",
   "neqo-transport",
-  "neqo-interop",
   "test-fixture",
 ]
 
@@ -16,7 +15,7 @@ homepage = "https://github.com/mozilla/neqo/"
 repository = "https://github.com/mozilla/neqo/"
 authors = ["The Neqo Authors <necko@mozilla.com>"]
 version = "0.7.1"
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0"
 # Don't increase beyond what Firefox is currently using:
 # https://firefox-source-docs.mozilla.org/writing-rust-code/update-policy.html#schedule
