@@ -617,7 +617,11 @@ impl<'a> Frame<'a> {
 mod tests {
     use neqo_common::{Decoder, Encoder};
 
-    use super::*;
+    use crate::{
+        cid::MAX_CONNECTION_ID_LEN,
+        frame::{AckRange, Frame, FRAME_TYPE_ACK},
+        CloseError, Error, StreamId, StreamType,
+    };
 
     fn just_dec(f: &Frame, s: &str) {
         let encoded = Encoder::from_hex(s);
